+++ spec/lumberjack/attributes_helper_spec.rb
@@ -2,9 +2,28 @@
 
 require "spec_helper"
 
-RSpec.describe Lumberjack::TagContext do
+RSpec.describe Lumberjack::AttributesHelper do
   let(:attributes) { {} }
-  let(:tag_context) { Lumberjack::TagContext.new(attributes) }
+  let(:tag_context) { Lumberjack::AttributesHelper.new(attributes) }
+
+  describe "expand_runtime_values" do
+    it "should return the identical hash as is if there are no Procs" do
+      hash = {"foo" => 1, "bar" => 2}
+      expect(Lumberjack::AttributesHelper.expand_runtime_values(hash)).to equal(hash)
+    end
+
+    it "should replace all keys with strings" do
+      hash = {foo: 1, bar: 2}
+      expect(Lumberjack::AttributesHelper.expand_runtime_values(hash)).to eq({"foo" => 1, "bar" => 2})
+    end
+
+    it "should replace Procs that take no arguments with the runtime value" do
+      p1 = lambda { "stuff" }
+      p2 = lambda { |x| x.upcase }
+      hash = {foo: 1, bar: p1, baz: p2}
+      expect(Lumberjack::AttributesHelper.expand_runtime_values(hash)).to eq({"foo" => 1, "bar" => "stuff", "baz" => p2})
+    end
+  end
 
   describe "#to_h" do
     it "returns a copy of the attributes" do